requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pymupdf==1.24.1
pillow==10.3.0
python-docx==1.1.0
openai==0.28.1
